extract the heuristic code
