+++ Backend/map/walkService.js
@@ -1,7 +1,8 @@
 let graph = require("./walkModel");
 let path = require("ngraph.path");
 const haversine = require("haversine");
-let manhattan = require("manhattan-distance");
+
+let { harversine_heuristic } = require("./heuristic");
 
 function checkGraph() {
   graph.forEachNode((node) => {
@@ -16,7 +17,7 @@ function findShortestPath(source, target) {
       return link.data.distance;
     },
     heuristic(fromNode, toNode) {
-      return normal_distance(fromNode, toNode);
+      return harversine_heuristic(fromNode, toNode);
     },
   });
 
@@ -24,27 +25,6 @@ function findShortestPath(source, target) {
   let shortestPath = pathFinder.find(source, target);
 
   return shortestPath;
-}
-
-//normal distance formula
-function normal_distance(fromNode, toNode) {
-  let dx = fromNode.data.coordinates[0] - toNode.data.coordinates[1];
-  let dy = fromNode.data.coordinates[0] - toNode.data.coordinates[1];
-
-  return Math.sqrt(dx * dx + dy * dy);
-}
-
-//wrosen the result
-function manhattan_heuristic(fromNode, toNode) {
-  //manhattan distance
-  return (
-    manhattan(
-      fromNode.data.coordinates[0],
-      fromNode.data.coordinates[1],
-      toNode.data.coordinates[0],
-      toNode.data.coordinates[1]
-    ).slice(0, -2) * 1000
-  );
 }
 
 /**
@@ -211,6 +191,7 @@ function calculateRequestPath(source, target, percentage, isMax) {
       path: nodes,
       elevationGain: shortestElevationGain,
       distance: shortestDistance,
+      shortestDistance: shortestDistance,
     };
   } else {
     //compute all the paths
